Started implementing item limit checking
- Adding in checks to prevent users selecting more magic/faction items in their
  army than are allowed in the items' limit (usually just one as items cannot
  typically be duplicated - however I added a limit field to item struct for
  edge cases).

- Slowly starting to introduce the newly formatted item budget entries in the roster
  with a new parsing method in roster_parser to handle these. Going through all the army
  rosters and changing over to this system will require a lot more careful work, probably
  won't have this (corresponding to Issue #2) fully implemented for quite some time due
  to this.

+++ src/item_parser.cpp
@@ -66,7 +66,7 @@ namespace tools {
                 std::vector<std::string> vec = tools::split(allowed, ',');
                 for (const auto& x : vec) itm.allowed_units.insert(x);
             }
-            if (limit != "None") itm.limit = std::stoi(limit);
+            if (limit != "None") itm.limit = static_cast<unsigned int>(std::stoul(limit));
             else itm.limit = std::numeric_limits<int>::max();
             itm.restrictions = restrs;
             items.push_back(itm);
